check if we are on the main branch, also if the branch is clean (#13)
* check if we are on the main branch, also if the branch is clean

+++ deploy.sh
@@ -6,4 +6,16 @@ if [ -f .env ]; then
     export $(cat .env | grep -v '#' | awk '/=/ {print $1}')
 fi
 
-rsync -e "/usr/bin/ssh" --verbose  --bwlimit=2000 -av ./dist/ $USER@$DOMAIN:/home/$USER/$DOMAIN/ --delete-after
+currentBranch="$(git symbolic-ref --short -q HEAD 2>/dev/null)" ||
+currentBranch="(unnamed branch)"
+
+if [ -z "$(git status --porcelain)" ]; then
+  echo ""
+else
+  echo "Branch is not clean"
+  exit
+fi
+
+if [[ $currentBranch -eq $EXPECTED_BRANCH ]]; then
+  rsync -e "/usr/bin/ssh" --verbose  --bwlimit=2000 -av ./dist/ $USER@$DOMAIN:/home/$USER/$DOMAIN/ --delete-after
+fi
